remove the unused deps
