--- cypress/jsconfig.json
@@ -1,3 +1,0 @@
-{
-  "include": ["./node_modules/cypress", "cypress/**/*.js"]
-}
