+++ cypress/support/commands.ts
@@ -1,6 +1,5 @@
-import { waitForGQL } from "../utils/networking";
+import { createUser } from "../utils/evergreen";
 
-const GQL_QUERY = "gqlQuery";
 const LOGIN_COOKIE = "mci-token";
 const TOAST_COOKIE = "announcement-toast";
 const loginURL = "http://localhost:9090/login";
@@ -38,27 +37,19 @@ Cypress.Commands.add("preserveCookies", () => {
   );
 });
 
-Cypress.Commands.add("listenGQL", () => {
-  cy.server();
-  cy.route("POST", "/graphql/query").as(GQL_QUERY);
-});
-
-Cypress.Commands.add("waitForGQL", (queryName, options = {}) =>
-  waitForGQL(`@${GQL_QUERY}`, queryName, options)
-);
-
-Cypress.Commands.add("dataCy", (value, options) =>
+type cyGetOptions = Parameters<typeof cy.get>[1];
+Cypress.Commands.add("dataCy", (value: string, options: cyGetOptions) =>
   cy.get(`[data-cy=${value}]`, options)
 );
-Cypress.Commands.add("dataRowKey", (value, options) =>
+Cypress.Commands.add("dataRowKey", (value: string, options: cyGetOptions) =>
   cy.get(`[data-row-key=${value}]`, options)
 );
 
-Cypress.Commands.add("dataTestId", (value, options) =>
+Cypress.Commands.add("dataTestId", (value: string, options: cyGetOptions) =>
   cy.get(`[data-test-id=${value}]`, options)
 );
 
-Cypress.Commands.add("getInputByLabel", (label) =>
+Cypress.Commands.add("getInputByLabel", (label: string) =>
   cy
     .contains("label", label)
     .invoke("attr", "for")
@@ -67,10 +58,15 @@ Cypress.Commands.add("getInputByLabel", (label) =>
     })
 );
 
-Cypress.Commands.add("toggleTableFilter", (colNum) => {
+Cypress.Commands.add("toggleTableFilter", (colNum: number) => {
   cy.get(`.ant-table-thead > tr > :nth-child(${colNum})`)
     .find("[role=button]")
     .first()
     .click();
   cy.get(".ant-table-filter-dropdown").should("be.visible");
+});
+
+type CreateUserParams = Parameters<typeof createUser>[0];
+Cypress.Commands.add("createUser", (options: CreateUserParams) => {
+  createUser(options);
 });
